changed terrain3D's export to work in other way to make it work in runtime
moved from using textures as data for all map. i had to do it because of
lack of support for .exr files in builded version of game(fuck godot)

now i just import region.res files, it's actually simpler.
but it was a lot of traial, pain and error to get it to work.